+++ tests/test_profile_general.py
@@ -125,6 +125,13 @@ ALL_CSVS_MI350 = sorted(
         "pmc_perf_5.csv",
         "pmc_perf_6.csv",
         "pmc_perf_7.csv",
+        "pmc_perf_8.csv",
+        "pmc_perf_9.csv",
+        "pmc_perf_10.csv",
+        "pmc_perf_11.csv",
+        "pmc_perf_12.csv",
+        "pmc_perf_13.csv",
+        "pmc_perf_14.csv",
         "sysinfo.csv",
     ]
 )
@@ -312,9 +319,10 @@ def gpu_soc():
 
     ## 3) Deduce gpu model name from arch
     gpu_model = list(SUPPORTED_ARCHS[gpu_arch].keys())[0].upper()
-    if gpu_model not in ("MI50", "MI100", "MI200"):
-        if chip_id in CHIP_IDS:
-            gpu_model = CHIP_IDS[chip_id]
+    # For testing purposes we only care about gpu model series not the specific model
+    # if gpu_model not in ("MI50", "MI100", "MI200"):
+    #     if chip_id in CHIP_IDS:
+    #         gpu_model = CHIP_IDS[chip_id]
 
     return gpu_model
 
@@ -322,15 +330,22 @@ def gpu_soc():
 soc = gpu_soc()
 
 # Set rocprofv2 as profiler if MI300
-if soc == "MI100":
-    os.environ["ROCPROF"] = "rocprof"
+if "ROCPROF" not in os.environ.keys():
+    if soc == "MI100":
+        os.environ["ROCPROF"] = "rocprof"
 
-else:
-    os.environ["ROCPROF"] = "rocprofv3"
+    else:
+        os.environ["ROCPROF"] = "rocprofv3"
 
 
 def using_v3():
-    return "ROCPROF" in os.environ.keys() and os.environ["ROCPROF"].endswith("rocprofv3")
+    return "ROCPROF" not in os.environ.keys() or (
+        "ROCPROF" in os.environ.keys()
+        and (
+            os.environ["ROCPROF"].endswith("rocprofv3")
+            or os.environ["ROCPROF"] == "rocprofiler-sdk"
+        )
+    )
 
 
 Baseline_dir = str(Path("tests/workloads/vcopy/" + soc).resolve())
@@ -542,7 +557,7 @@ def test_path(binary_handler_profile_rocprof_compute):
 
 @pytest.mark.misc
 def test_roof_kernel_names(binary_handler_profile_rocprof_compute):
-    if soc in ("MI100", "MI350"):
+    if soc in ("MI100"):
         # roofline is not supported on MI100
         assert True
         # Do not continue testing
@@ -557,7 +572,9 @@ def test_roof_kernel_names(binary_handler_profile_rocprof_compute):
     # assert successful run
     assert returncode == 0
     file_dict = test_utils.check_csv_files(workload_dir, 1, num_kernels)
-    if soc == "MI200" in soc or "MI300" in soc:
+    if soc == "MI100":
+        assert sorted(list(file_dict.keys())) == ALL_CSVS_MI100
+    else:
         assert sorted(list(file_dict.keys())) == sorted(
             (
                 [f for f in ROOF_ONLY_FILES if f != "timestamps.csv"]
@@ -566,8 +583,6 @@ def test_roof_kernel_names(binary_handler_profile_rocprof_compute):
             )
             + ["kernelName_legend.pdf"]
         )
-    else:
-        assert sorted(list(file_dict.keys())) == ALL_CSVS_MI100
 
     validate(
         inspect.stack()[0][3],
@@ -725,6 +740,7 @@ def test_block_SQ(binary_handler_profile_rocprof_compute):
             "pmc_perf_5.csv",
             "pmc_perf_6.csv",
             "pmc_perf_7.csv",
+            "pmc_perf_8.csv",
             "sysinfo.csv",
         ]
 
@@ -754,7 +770,7 @@ def test_block_SQC(binary_handler_profile_rocprof_compute):
         "sysinfo.csv",
         "timestamps.csv",
     ]
-    if using_v3() or soc == "MI350":
+    if using_v3():
         expected_csvs.remove("timestamps.csv")
 
     assert sorted(list(file_dict.keys())) == sorted(expected_csvs)
@@ -788,8 +804,25 @@ def test_block_TA(binary_handler_profile_rocprof_compute):
         "sysinfo.csv",
         "timestamps.csv",
     ]
-    if using_v3() or soc == "MI350":
+
+    if using_v3():
         expected_csvs.remove("timestamps.csv")
+
+    if soc == "MI350":
+        expected_csvs = [
+            "pmc_perf.csv",
+            "pmc_perf_0.csv",
+            "pmc_perf_1.csv",
+            "pmc_perf_2.csv",
+            "pmc_perf_3.csv",
+            "pmc_perf_4.csv",
+            "pmc_perf_5.csv",
+            "pmc_perf_6.csv",
+            "pmc_perf_7.csv",
+            "pmc_perf_8.csv",
+            "pmc_perf_9.csv",
+            "sysinfo.csv",
+        ]
 
     assert sorted(list(file_dict.keys())) == sorted(expected_csvs)
 
@@ -894,6 +927,26 @@ def test_block_TCP(binary_handler_profile_rocprof_compute):
             "sysinfo.csv",
         ]
 
+    if soc == "MI350":
+        expected_csvs = [
+            "pmc_perf.csv",
+            "pmc_perf_0.csv",
+            "pmc_perf_1.csv",
+            "pmc_perf_2.csv",
+            "pmc_perf_3.csv",
+            "pmc_perf_4.csv",
+            "pmc_perf_5.csv",
+            "pmc_perf_6.csv",
+            "pmc_perf_7.csv",
+            "pmc_perf_8.csv",
+            "pmc_perf_9.csv",
+            "pmc_perf_10.csv",
+            "pmc_perf_11.csv",
+            "pmc_perf_12.csv",
+            "pmc_perf_13.csv",
+            "sysinfo.csv",
+        ]
+
     if soc == "MI100" or soc == "MI200":
         expected_csvs = [
             "pmc_perf.csv",
@@ -926,9 +979,6 @@ def test_block_TCP(binary_handler_profile_rocprof_compute):
                 "pmc_perf_9.csv",
                 "sysinfo.csv",
             ]
-
-    if soc == "MI350":
-        expected_csvs.remove("timestamps.csv")
 
     assert sorted(list(file_dict.keys())) == sorted(expected_csvs)
 
@@ -984,6 +1034,32 @@ def test_block_TCC(binary_handler_profile_rocprof_compute):
             "sysinfo.csv",
         ]
 
+    if soc == "MI350":
+        expected_csvs = [
+            "pmc_perf.csv",
+            "pmc_perf_0.csv",
+            "pmc_perf_1.csv",
+            "pmc_perf_2.csv",
+            "pmc_perf_3.csv",
+            "pmc_perf_4.csv",
+            "pmc_perf_5.csv",
+            "pmc_perf_6.csv",
+            "pmc_perf_7.csv",
+            "pmc_perf_8.csv",
+            "pmc_perf_9.csv",
+            "pmc_perf_10.csv",
+            "pmc_perf_11.csv",
+            "pmc_perf_12.csv",
+            "pmc_perf_13.csv",
+            "pmc_perf_14.csv",
+            "pmc_perf_15.csv",
+            "pmc_perf_16.csv",
+            "pmc_perf_17.csv",
+            "pmc_perf_18.csv",
+            "pmc_perf_19.csv",
+            "sysinfo.csv",
+        ]
+
     if soc == "MI100" or soc == "MI200":
         expected_csvs = [
             "pmc_perf.csv",
@@ -1018,9 +1094,6 @@ def test_block_TCC(binary_handler_profile_rocprof_compute):
                 "pmc_perf_10.csv",
                 "sysinfo.csv",
             ]
-
-    if soc == "MI350":
-        expected_csvs.remove("timestamps.csv")
 
     assert sorted(list(file_dict.keys())) == sorted(expected_csvs)
 
@@ -1083,6 +1156,7 @@ def test_block_SPI(binary_handler_profile_rocprof_compute):
             "pmc_perf_9.csv",
             "pmc_perf_10.csv",
             "pmc_perf_11.csv",
+            "pmc_perf_12.csv",
             "sysinfo.csv",
         ]
 
@@ -1163,7 +1237,7 @@ def test_block_CPF(binary_handler_profile_rocprof_compute):
         "sysinfo.csv",
         "timestamps.csv",
     ]
-    if using_v3() or soc == "MI350":
+    if using_v3():
         expected_csvs.remove("timestamps.csv")
     assert sorted(list(file_dict.keys())) == sorted(expected_csvs)
 
@@ -1249,6 +1323,7 @@ def test_block_SQ_CPC(binary_handler_profile_rocprof_compute):
             "pmc_perf_5.csv",
             "pmc_perf_6.csv",
             "pmc_perf_7.csv",
+            "pmc_perf_8.csv",
             "sysinfo.csv",
         ]
 
@@ -1336,6 +1411,7 @@ def test_block_SQ_TA(binary_handler_profile_rocprof_compute):
             "pmc_perf_5.csv",
             "pmc_perf_6.csv",
             "pmc_perf_7.csv",
+            "pmc_perf_8.csv",
             "sysinfo.csv",
         ]
 
@@ -1419,6 +1495,7 @@ def test_block_SQ_SPI(binary_handler_profile_rocprof_compute):
             "pmc_perf_5.csv",
             "pmc_perf_6.csv",
             "pmc_perf_7.csv",
+            "pmc_perf_8.csv",
             "sysinfo.csv",
         ]
 
@@ -1507,6 +1584,7 @@ def test_block_SQ_SQC_TCP_CPC(binary_handler_profile_rocprof_compute):
             "pmc_perf_5.csv",
             "pmc_perf_6.csv",
             "pmc_perf_7.csv",
+            "pmc_perf_8.csv",
             "sysinfo.csv",
         ]
 
@@ -1609,6 +1687,13 @@ def test_block_SQ_SPI_TA_TCC_CPF(binary_handler_profile_rocprof_compute):
             "pmc_perf_5.csv",
             "pmc_perf_6.csv",
             "pmc_perf_7.csv",
+            "pmc_perf_8.csv",
+            "pmc_perf_9.csv",
+            "pmc_perf_10.csv",
+            "pmc_perf_11.csv",
+            "pmc_perf_12.csv",
+            "pmc_perf_13.csv",
+            "pmc_perf_14.csv",
             "sysinfo.csv",
         ]
 
@@ -1816,7 +1901,7 @@ def test_join_type_kernel(binary_handler_profile_rocprof_compute):
 @pytest.mark.sort
 def test_roof_sort_dispatches(binary_handler_profile_rocprof_compute):
     # only test 1 device for roofline
-    if soc in ("MI100", "MI350"):
+    if soc in ("MI100"):
         # roofline is not supported on MI100
         assert True
         # Do not continue testing
@@ -1851,7 +1936,7 @@ def test_roof_sort_dispatches(binary_handler_profile_rocprof_compute):
 @pytest.mark.sort
 def test_roof_sort_kernels(binary_handler_profile_rocprof_compute):
     # only test 1 device for roofline
-    if soc in ("MI100", "MI350"):
+    if soc in ("MI100"):
         # roofline is not supported on MI100
         assert True
         # Do not continue testing
@@ -1886,7 +1971,7 @@ def test_roof_sort_kernels(binary_handler_profile_rocprof_compute):
 @pytest.mark.mem
 def test_roof_mem_levels_vL1D(binary_handler_profile_rocprof_compute):
     # only test 1 device for roofline
-    if soc in ("MI100", "MI350"):
+    if soc in ("MI100"):
         # roofline is not supported on MI100
         assert True
         # Do not continue testing
@@ -1921,7 +2006,7 @@ def test_roof_mem_levels_vL1D(binary_handler_profile_rocprof_compute):
 @pytest.mark.mem
 def test_roof_mem_levels_LDS(binary_handler_profile_rocprof_compute):
     # only test 1 device for roofline
-    if soc in ("MI100", "MI350"):
+    if soc in ("MI100"):
         # roofline is not supported on MI100
         assert True
         # Do not continue testing